Parse elastic-2.0 style error objects
ES now hides the string that we're expecting
in the root_cause object so they can include
a bunch more error metadata. If the root_cause
object is present, assume it's a 2.0 error and
extract the string from there, otherwise assume it's
a 1.x error and proceed as in Jut 1.0.

I don't have a way to consistently reproduce this situation
so there's no test, but this is what Orestes does to
catch the error where you don't have aggkey.groovy so
presumably it works a little.