read/write -timeField
Now the point that the adapter will use as a timestamp for documents in
ES is configurable by the timeField option. Juttle points still have
"time", this just determines the name of the field that the adapter will
put pt.time in when writing and put into pt.time when reading.