optimizer: build date histograms with timeField The date histogram builder assumes that the time field is always 'time'. That's not right. It should use the timeField parameter.
This defeated our tests because we always write points with a "time" field
as well as the timeField, so date histograms on the "time" field work
even though it's not the timeField. Let's drop the "time" field from points
with a timeField on write.

fixes https://github.com/juttle/juttle-elastic-adapter/issues/66