read: don't apply limit globally
read elastic uses the same logic for limit as it does for head and
tail, so across any number of fetches it won't return more than a
total of limit points. This is not the right way to enforce limit,
which only applies to each individual call to the read() function.

Fixes https://github.com/juttle/juttle-elastic-adapter/issues/115
Fixes https://github.com/juttle/juttle-elastic-adapter/issues/116